Fix paths to modules. Last commit config and accountmanager were moved to a lib, but the path has not been replaced.
