+++ libturpial/lib/accountmanager.py
@@ -8,7 +8,7 @@
 import logging
 
 from libturpial.common import LoginStatus
-from libturpial.config import AccountConfig
+from libturpial.lib.config import AccountConfig
 from libturpial.api.models.account import Account
